+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-* Upgrade the `org.realityforge.proton` artifact to version `0.06`.
+* Upgrade the `org.realityforge.proton` artifact to version `0.07`.
 
 ### [v0.161](https://github.com/arez/arez/tree/v0.161) (2019-12-25) · [Full Changelog](https://github.com/arez/arez/compare/v0.160...v0.161)